+++ src/routes/v1/search.ts
@@ -5,14 +5,18 @@ const router = new Router({
   prefix: '/search'
 })
 
-/**
- * Build elasticsearch queries, filter out duplicates, and send back to client
- */
 router.get('/', async (ctx) => {
   const searchText = (ctx.query.q as string)?.trim()
   const taxonomies = (ctx.query.taxonomies as string)?.trim()
   const results = await searchService.search({ searchText, taxonomies })
   ctx.body = results
+})
+
+router.get('/filters', async (ctx) => {
+  const searchText = (ctx.query.query as string)?.trim()
+  const taxonomies = (ctx.query.taxonomies as string)?.trim()
+  const filters = await searchService.getFilters(searchText, taxonomies)
+  ctx.body = filters
 })
 
 export default router
