+++ src/util.ts
@@ -32,3 +32,25 @@ export function getRandomString(len: number = 30, characters: 'ALPHANUMERIC' | '
 export function decodeToken(token: string) {
   return Buffer.from(token, 'base64').toString('utf-8')
 }
+
+export function timeStringToDate(time: string) {
+  const re = /(\d+):(\d+)\s?(am|pm)/i
+  const match = re.exec(time)
+  if (!match) {
+    return null
+  }
+  let h = Number(match[1])
+  const m = Number(match[2])
+  const z = match[3].toLowerCase()
+  if (z === 'am' && h === 12) {
+    h = 0
+  } else if (z === 'pm' && h < 12) {
+    h += 12
+  }
+  const date = new Date()
+  const Y = date.getFullYear()
+  const M = date.getMonth()
+  const D = date.getDate()
+  const result = new Date(Y, M, D, h, m)
+  return result
+}
